+++ app/api/user/config/route.ts
@@ -5,18 +5,23 @@
  * so they persist across browsers, devices, and PWA installs.
  */
 
-import { Redis } from '@upstash/redis/cloudflare';
 import { NextRequest, NextResponse } from 'next/server';
 import { authenticationRequiredResponse } from '@/lib/server/api-responses';
 import { getServerSession } from '@/lib/server/auth';
+import { getRedisClient } from '@/lib/server/redis';
 
 export const runtime = 'edge';
-
-const redis = Redis.fromEnv();
 
 function redisKey(profileId: string): string {
   const safe = profileId.replace(/[^a-zA-Z0-9_-]/g, '');
   return `user:config:${safe}`;
+}
+
+function syncUnavailableResponse() {
+  return NextResponse.json(
+    { error: 'Server-side sync is not configured on this deployment' },
+    { status: 503 }
+  );
 }
 
 export async function GET(request: NextRequest) {
@@ -25,6 +30,11 @@ export async function GET(request: NextRequest) {
 
   if (!profileId) {
     return authenticationRequiredResponse();
+  }
+
+  const redis = getRedisClient();
+  if (!redis) {
+    return syncUnavailableResponse();
   }
 
   try {
@@ -45,6 +55,11 @@ export async function POST(request: NextRequest) {
 
   if (!profileId) {
     return authenticationRequiredResponse();
+  }
+
+  const redis = getRedisClient();
+  if (!redis) {
+    return syncUnavailableResponse();
   }
 
   try {
