+++ app/api/user/sync/route.ts
@@ -1,19 +1,29 @@
-import { Redis } from '@upstash/redis/cloudflare';
 import { NextRequest, NextResponse } from 'next/server';
 import { authenticationRequiredResponse } from '@/lib/server/api-responses';
 import { getServerSession } from '@/lib/server/auth';
+import { getRedisClient } from '@/lib/server/redis';
 
 // 确保这行代码在整个文件中只出现一次
 export const runtime = 'edge';
 
-const redis = Redis.fromEnv();
+function syncUnavailableResponse() {
+  return NextResponse.json(
+    { error: 'Server-side sync is not configured on this deployment' },
+    { status: 503 }
+  );
+}
 
 export async function GET(request: NextRequest) {
   const session = await getServerSession(request);
   const profileId = session?.profileId;
-  
+
   if (!profileId) {
     return authenticationRequiredResponse();
+  }
+
+  const redis = getRedisClient();
+  if (!redis) {
+    return syncUnavailableResponse();
   }
 
   try {
@@ -31,9 +41,14 @@ export async function GET(request: NextRequest) {
 export async function POST(request: NextRequest) {
   const session = await getServerSession(request);
   const profileId = session?.profileId;
-  
+
   if (!profileId) {
     return authenticationRequiredResponse();
+  }
+
+  const redis = getRedisClient();
+  if (!redis) {
+    return syncUnavailableResponse();
   }
 
   try {
