+++ lib/server/auth.ts
@@ -1,6 +1,6 @@
-import { Redis } from '@upstash/redis/cloudflare';
-import { getOptionalRequestContext } from '@cloudflare/next-on-pages';
 import { NextRequest, NextResponse } from 'next/server';
+import { getRedisClient } from '@/lib/server/redis';
+import { getRuntimeEnvValue } from '@/lib/server/runtime-env';
 import { getRuntimeFeatures } from '@/lib/server/runtime-features';
 import {
   createStoredAccount,
@@ -87,49 +87,16 @@ const DANMAKU_API_URL = process.env.DANMAKU_API_URL || process.env.NEXT_PUBLIC_D
 const SESSION_MAX_AGE_SECONDS = 60 * 60 * 24 * 30;
 const MANAGED_AUTH_FORCED = process.env.MANAGED_AUTH_ENABLED === 'true';
 
-function getRuntimeEnvValue(name: string, fallback = ''): string {
-  try {
-    const runtimeEnv = getOptionalRequestContext()?.env as unknown as Record<string, unknown> | undefined;
-    const value = runtimeEnv?.[name];
-    if (typeof value === 'string') return value;
-  } catch {
-    // Outside Cloudflare's request runtime, fall back to process.env.
-  }
-
-  return process.env[name] || fallback;
-}
-
 function getEffectiveAdminPassword(): string {
   return getRuntimeEnvValue('ADMIN_PASSWORD', ADMIN_PASSWORD) ||
     getRuntimeEnvValue('ACCESS_PASSWORD', ACCESS_PASSWORD);
 }
-
-let cachedRedis: Redis | null | undefined;
 
 export class ManagedAuthStorageError extends Error {
   constructor(operation: 'read' | 'write', cause?: unknown) {
     super(`Managed auth storage ${operation} failed`, { cause });
     this.name = 'ManagedAuthStorageError';
   }
-}
-
-function getRedisClient(): Redis | null {
-  if (cachedRedis !== undefined) {
-    return cachedRedis;
-  }
-
-  const url = getRuntimeEnvValue('UPSTASH_REDIS_REST_URL');
-  const token = getRuntimeEnvValue('UPSTASH_REDIS_REST_TOKEN');
-  if (!url || !token) {
-    cachedRedis = null;
-    return cachedRedis;
-  }
-
-  cachedRedis = new Redis({
-    url,
-    token,
-  });
-  return cachedRedis;
 }
 
 function isManagedAuthEnabled(): boolean {
